Adding parameterized Type to IR
It's currently called TypeVar, which is not a great name. Like the
recent changes to Type, this can now hold the Type's parameter as a
variable. Since the text format can allow variables to be indexes or
strings, this uses Var. None of the functionality is actually used yet,
but TypeVar is correctly forwarded through the code and can pass all the
tests.